// apps/frontend/src/api/model/listPagesType.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Workspace Management API
 * API specification for managing users, workspaces, and pages.
 * OpenAPI spec version: 1.0.0
 */

export type ListPagesType = typeof ListPagesType[keyof typeof ListPagesType];


export const ListPagesType = {
  course: 'course',
  assignment: 'assignment',
  folder: 'folder',
  note: 'note',
} as const;